+++ packages/core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @usedapp/core
 
+## 1.2.16
+
+### Patch Changes
+
+- fb44f98: 🔬 Make useRawLogs more efficient and responsive
+
+## 1.2.15
+
+### Patch Changes
+
+- c688928: 🌝 Add optimism sepolia to chain ids
+
+## 1.2.14
+
+### Patch Changes
+
+- 1ae0bcb: Added 'Optimism Sepolia' testnet
+
 ## 1.2.13
 
 ### Patch Changes